+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Add faces to sources that use files.
 * Add faces to `helm-projectile-browse-dirty-projects`.
 * Add autoloads to `helm-projectile` commands defined with `helm-projectile-command`.
+* Make `helm-projectile-ag` to use common ignore list
 
 ## 1.2.0 (2025-07-21)